// GridCraftCdk/grid-craft-angular/src/app/shared/apis/generated-apis.ts
export class GridColumnInput implements IGridColumnInput {
    name!: string | undefined;
    expression!: string | undefined;

    constructor(data?: IGridColumnInput) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.expression = _data["expression"];
        }
    }

    static fromJS(data: any): GridColumnInput {
        data = typeof data === 'object' ? data : {};
        let result = new GridColumnInput();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["expression"] = this.expression;
        return data;
    }
}

export interface IGridColumnInput {
    name: string | undefined;
    expression: string | undefined;
}
// ...
export interface IGridInput {
    numberOfRows: number;
    columns: GridColumnInput[] | undefined;
}